Adicionada função de rota!

+++ src/main/java/br/com/agenda/entidade/Rota.java
@@ -17,18 +17,18 @@ public class Rota implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
-	
+	private Long id;
+
 	private String descricao;
-	
+
 	@OneToMany(cascade = CascadeType.ALL, mappedBy = "rota")
 	private List<Cidade> cidades;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
